+++ lizard_damage/tasks.py
@@ -37,7 +37,7 @@ def benefit_scenario_to_task(benefit_scenario, username="admin"):
     """
     Send provided benefit scenario as task
     """
-    task_name = 'Scenario ({}) calculate benefit'.format(benefit_scenario.id)
+    task_name = 'Scenario (%05d) calculate benefit' % benefit_scenario.id
     task_kwargs = (
         '{"username": "%s", "taskname": "%s", "benefit_scenario_id": "%d"}' % (
             username, task_name, benefit_scenario.id))