+++ CHANGES.rst
@@ -7,6 +7,8 @@ Changelog of lizard-damage
 
 - Update cleanup command.
 
+- Fix names of tasks to be picked up by cleanup command.
+
 
 2.0.1 (2015-03-26)
 ------------------